requirements.txt
python-fasthtml>=0.6.9
edge-tts>=6.1.9
pypinyin>=0.50.0
requests>=2.28.0
python-dotenv>=1.0.0
jieba>=0.42.1
openai>=1.0.0
opencc-python-reimplemented>=1.7.0